+++ src/modules/TaskBoard/components/taskList/models/sortTasksByPriority.ts
@@ -1,4 +1,4 @@
-import { getHighestPriority } from '../components/Tags/model/getHighestPriority'
+import { getHighestPriority } from '@/features/tags/model/getHighestPriority'
 import { taskModel } from '@/modules/TaskBoard/model/task'
 
 /**
